Fix broken SetupActivity compose-hierarchy tests from commit c5b75f3
CI on commit c5b75f3 failed PermissionsDeniedE2ETest#setupFlow_reachesMediaStep_whenPermissionNotGranted:

  java.lang.IllegalStateException: No compose hierarchies found in the app.
    at androidx.compose.ui.test.AssertionsKt.assertIsDisplayed(Assertions.kt:33)

createEmptyComposeRule() combined with a manually launched
ActivityScenario.launch(SetupActivity::class.java) does not reliably
let Compose's test framework discover SetupActivity's compose
hierarchy. Worse, the mirrored test in PermissionsGrantedE2ETest
(asserting the step is *absent* via assertDoesNotExist()) had the
identical underlying problem but never surfaced it in CI:
assertDoesNotExist() fetches nodes with atLeastOneRootRequired =
false, so it passes vacuously when no hierarchy is found at all,
rather than actually confirming the step is absent. Both new tests
were unreliable; only one visibly failed.

Fix: move both assertions into two new dedicated files,
SetupActivityDeniedE2ETest and SetupActivityGrantedE2ETest, using
createAndroidComposeRule<SetupActivity>() -- the same class-level rule
SetupActivityTest already relies on successfully across every prior
round of this PR. The rule launches SetupActivity itself and is
guaranteed to be wired into Compose's test synchronization before any
assertion runs, unlike a manually launched ActivityScenario alongside
an unrelated empty rule. Removed the broken versions from
PermissionsGrantedE2ETest/PermissionsDeniedE2ETest, restoring those
two files to their previously-reviewed, already-passing state.

Wired two new CI steps mirroring the existing PermissionsGrantedE2ETest
/ PermissionsDeniedE2ETest pattern (including the same
-PmediaPermissionGranted=false precondition for the denied side and a
re-grant step afterward), folded into both --outcome ternaries in
"Write test-result summary" and "Gate on test failures".

🦀
