+++ .github/workflows/build.yml
@@ -622,6 +622,93 @@ jobs:
           path: results/e2e-permissions-denied.ndjson
           retention-days: 14
 
+      - name: Run SetupActivityGrantedE2ETest
+        # No `continue-on-error` (issue #309): record the real outcome and defer
+        # the pass/fail verdict to the "Gate on test failures" step.
+        #
+        # Issue #509/#566 follow-up (PR #564 review): an earlier attempt folded this
+        # assertion into PermissionsGrantedE2ETest using createEmptyComposeRule() plus
+        # a manually launched ActivityScenario for SetupActivity. That combination let
+        # this exact test pass, but only because assertDoesNotExist() vacuously
+        # succeeds when Compose's test framework finds no hierarchy at all to search
+        # (see SetupActivityGrantedE2ETest's class doc). This dedicated step uses
+        # createAndroidComposeRule<SetupActivity>() instead, the same rule
+        # SetupActivityTest already relies on successfully, so the hierarchy is
+        # actually found and the assertion is meaningful.
+        id: run_setup_activity_granted_e2e
+        if: steps.diff_check.outputs.needs_full_build == 'true'
+        run: |
+          ADB="$ANDROID_HOME/platform-tools/adb"
+          "$ADB" shell input keyevent KEYCODE_WAKEUP
+          "$ADB" shell input swipe 300 1000 300 300
+          sleep 2
+          mkdir -p results
+          set -o pipefail
+          LOGCAT_SINCE=$("$ADB" shell "date '+%m-%d %H:%M:%S.000'" | tr -d '\r')
+          if ./gradlew connectedE2EAndroidTest \
+            -Pe2eClass=com.gb4pc.e2e.SetupActivityGrantedE2ETest \
+            | tee >(grep '^##GB4PC_TEST##' > results/e2e-setup-activity-granted.ndjson); then
+            echo "outcome=success" >> "$GITHUB_OUTPUT"
+          else
+            echo "outcome=failure" >> "$GITHUB_OUTPUT"
+            echo "=== LOGCAT (since test start) ==="
+            "$ADB" logcat -d -T "$LOGCAT_SINCE" | \
+              bash scripts/filter_logcat.sh
+          fi
+
+      - name: Upload SetupActivityGrantedE2ETest markers
+        if: always() && steps.diff_check.outputs.needs_full_build == 'true'
+        uses: actions/upload-artifact@v7
+        with:
+          name: testresults-e2e-setup-activity-granted
+          path: results/e2e-setup-activity-granted.ndjson
+          retention-days: 14
+
+      - name: Run SetupActivityDeniedE2ETest
+        # No `continue-on-error` (issue #309): record the real outcome and defer
+        # the pass/fail verdict to the "Gate on test failures" step.
+        #
+        # -PmediaPermissionGranted=false, same as PermissionsDeniedE2ETest above and for
+        # the same reason: this class asserts the permission is NOT granted, so it must
+        # never run under the task's default grant.
+        id: run_setup_activity_denied_e2e
+        if: steps.diff_check.outputs.needs_full_build == 'true'
+        run: |
+          ADB="$ANDROID_HOME/platform-tools/adb"
+          "$ADB" shell input keyevent KEYCODE_WAKEUP
+          "$ADB" shell input swipe 300 1000 300 300
+          sleep 2
+          mkdir -p results
+          set -o pipefail
+          LOGCAT_SINCE=$("$ADB" shell "date '+%m-%d %H:%M:%S.000'" | tr -d '\r')
+          if ./gradlew connectedE2EAndroidTest \
+            -Pe2eClass=com.gb4pc.e2e.SetupActivityDeniedE2ETest \
+            -PmediaPermissionGranted=false \
+            | tee >(grep '^##GB4PC_TEST##' > results/e2e-setup-activity-denied.ndjson); then
+            echo "outcome=success" >> "$GITHUB_OUTPUT"
+          else
+            echo "outcome=failure" >> "$GITHUB_OUTPUT"
+            echo "=== LOGCAT (since test start) ==="
+            "$ADB" logcat -d -T "$LOGCAT_SINCE" | \
+              bash scripts/filter_logcat.sh
+          fi
+
+      - name: Re-grant media permission after SetupActivityDeniedE2ETest
+        # Mirrors the "Re-grant media permission after PermissionsDeniedE2ETest" step
+        # above: keeps device state in the "granted" baseline every other suite (and
+        # any future step) assumes.
+        if: always() && steps.diff_check.outputs.needs_full_build == 'true'
+        run: |
+          "$ANDROID_HOME/platform-tools/adb" shell pm grant com.gb4pc android.permission.READ_MEDIA_IMAGES
+
+      - name: Upload SetupActivityDeniedE2ETest markers
+        if: always() && steps.diff_check.outputs.needs_full_build == 'true'
+        uses: actions/upload-artifact@v7
+        with:
+          name: testresults-e2e-setup-activity-denied
+          path: results/e2e-setup-activity-denied.ndjson
+          retention-days: 14
+
       - name: Copy E2E test results
         if: always() && steps.diff_check.outputs.needs_full_build == 'true'
         continue-on-error: true
@@ -692,7 +779,7 @@ jobs:
               --outcome "${{ steps.run_instrumented.outcome }}" \
             app/build/outputs/androidTest-results/e2e \
               --suite-label "E2E Tests" \
-              --outcome "${{ (steps.run_overlay_e2e.outcome == 'skipped' && steps.run_gallery_e2e.outcome == 'skipped' && steps.run_permissions_granted_e2e.outcome == 'skipped' && steps.run_permissions_denied_e2e.outcome == 'skipped') && 'skipped' || ((steps.run_overlay_e2e.outcome == 'failure' || steps.run_gallery_e2e.outcome == 'failure' || steps.run_permissions_granted_e2e.outcome == 'failure' || steps.run_permissions_denied_e2e.outcome == 'failure') && 'failure' || 'success') }}"
+              --outcome "${{ (steps.run_overlay_e2e.outcome == 'skipped' && steps.run_gallery_e2e.outcome == 'skipped' && steps.run_permissions_granted_e2e.outcome == 'skipped' && steps.run_permissions_denied_e2e.outcome == 'skipped' && steps.run_setup_activity_granted_e2e.outcome == 'skipped' && steps.run_setup_activity_denied_e2e.outcome == 'skipped') && 'skipped' || ((steps.run_overlay_e2e.outcome == 'failure' || steps.run_gallery_e2e.outcome == 'failure' || steps.run_permissions_granted_e2e.outcome == 'failure' || steps.run_permissions_denied_e2e.outcome == 'failure' || steps.run_setup_activity_granted_e2e.outcome == 'failure' || steps.run_setup_activity_denied_e2e.outcome == 'failure') && 'failure' || 'success') }}"
 
       # Single authority on whether the test phase passed (issue #309). Replaces
       # the old blanket `continue-on-error` on the test steps: any failing test
@@ -714,7 +801,7 @@ jobs:
               --outcome "${{ steps.run_instrumented.outputs.outcome }}" \
             app/build/outputs/androidTest-results/e2e \
               --suite-label "E2E Tests" \
-              --outcome "${{ (steps.run_overlay_e2e.outputs.outcome == 'failure' || steps.run_gallery_e2e.outputs.outcome == 'failure' || steps.run_permissions_granted_e2e.outputs.outcome == 'failure' || steps.run_permissions_denied_e2e.outputs.outcome == 'failure') && 'failure' || 'success' }}"
+              --outcome "${{ (steps.run_overlay_e2e.outputs.outcome == 'failure' || steps.run_gallery_e2e.outputs.outcome == 'failure' || steps.run_permissions_granted_e2e.outputs.outcome == 'failure' || steps.run_permissions_denied_e2e.outputs.outcome == 'failure' || steps.run_setup_activity_granted_e2e.outputs.outcome == 'failure' || steps.run_setup_activity_denied_e2e.outputs.outcome == 'failure') && 'failure' || 'success' }}"
 
   file-issues:
     needs: [build-and-test]
